+++ packages/react-cookie/src/BrowserCookieManager.ts
@@ -1,15 +1,10 @@
 import cookie from 'cookie';
-import {hasDocumentCookie} from './utilities';
 import {CookieValue, Cookie} from './types';
 
 export {CookieContext} from './context';
 
 export class BrowserCookieManager {
   getCookie(name: string) {
-    if (!hasDocumentCookie()) {
-      return;
-    }
-
     const cookies = this.parse(document.cookie);
 
     return cookies[name] && cookies[name].value !== ''
@@ -18,20 +13,12 @@ export class BrowserCookieManager {
   }
 
   getCookies() {
-    if (!hasDocumentCookie()) {
-      return {};
-    }
-
     const cookies = this.parse(document.cookie || '');
 
     return cookies;
   }
 
   setCookie(name: string, value: string | CookieValue) {
-    if (!hasDocumentCookie()) {
-      return;
-    }
-
     const fullCookie = typeof value === 'string' ? {value} : value;
     const {value: cookieValue, ...options} = fullCookie;
 
@@ -39,27 +26,17 @@ export class BrowserCookieManager {
   }
 
   removeCookie(name: string) {
-    if (!hasDocumentCookie()) {
-      return;
-    }
-
     document.cookie = cookie.serialize(name, '', {
       expires: new Date('Thu, 01 Jan 1970 00:00:01 GMT'),
     });
   }
 
-  private parse(cookies: string) {
-    let newCookies: Cookie | undefined;
+  private parse(cookies: string): Cookie {
     const parsedCookies = cookie.parse(cookies);
 
-    Object.entries(parsedCookies).forEach(([key, value]) => {
-      if (newCookies) {
-        newCookies[key] = {value};
-      } else {
-        newCookies = {[key]: {value}};
-      }
-    });
-
-    return newCookies || {};
+    return Object.entries(parsedCookies).reduce(
+      (newCookies, [key, value]) => ({...newCookies, [key]: {value}}),
+      {} as Cookie,
+    );
   }
 }